+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Semantic ID for form answers
 - Add version number to distinguish question sets
 - Add Freetext question type for forms
+- Add Integer question with min and max values
 
 ### Changed
 - Increased SMS limit to 459 characters